+++ src/editor/utils.rs
@@ -1,5 +1,5 @@
 #[derive(Copy, Clone)]
 pub struct Size {
-    pub rows: u16,
-    pub cols: u16,
+    pub rows: usize,
+    pub cols: usize,
 }
